1、增加在线日志模块功能，上报查找设备失败的日志信息； 2、优化查找设备接口，增加连接设备的过滤规则。

+++ Swing/C/Login/SelectWatchViewController.m
@@ -39,7 +39,7 @@
 }
 
 - (void)backAction {
-    int count = self.navigationController.viewControllers.count;
+    NSUInteger count = self.navigationController.viewControllers.count;
     if(count > 2) {
         [self.navigationController popToViewController:self.navigationController.viewControllers[count - 3] animated:YES];
     }
